+++ zellij-utils/src/errors.rs
@@ -14,6 +14,7 @@ use colored::*;
 use log::error;
 use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Error, Formatter};
+use std::path::PathBuf;
 
 use miette::Diagnostic;
 use thiserror::Error as ThisError;
@@ -92,9 +93,16 @@ pub trait LoggableError<T>: Sized {
     ///     .print_error(|msg| println!("{msg}"))
     ///     .unwrap();
     /// ```
+    #[track_caller]
     fn print_error<F: Fn(&str)>(self, fun: F) -> Self;
 
     /// Convenienve function, calls `print_error` with the closure `|msg| log::error!("{}", msg)`.
+    // Dev note:
+    // Currently this hides the location of the caller, because it will show this very line as
+    // "source" of the logging call. This isn't correct, because it may have been called by other
+    // functions, too. To track this, we need to attach `#[track_caller]` to the closure below,
+    // which isn't stabilized yet: https://github.com/rust-lang/rust/issues/87417
+    #[track_caller]
     fn to_log(self) -> Self {
         self.print_error(|msg| log::error!("{}", msg))
     }
@@ -133,6 +141,7 @@ pub trait FatalError<T> {
     /// Discards the result type afterwards.
     ///
     /// [`to_log`]: LoggableError::to_log
+    #[track_caller]
     fn non_fatal(self);
 
     /// Mark results as being fatal.
@@ -393,6 +402,56 @@ pub enum ZellijError {
 
     #[error("failed to start PTY")]
     FailedToStartPty,
+
+    #[error(
+        "This version of zellij was built to load the core plugins from
+the globally configured plugin directory. However, a plugin wasn't found:
+
+    Plugin name: '{plugin_path}'
+    Plugin directory: '{plugin_dir}'
+
+If you're a user:
+    Please report this error to the distributor of your current zellij version
+
+If you're a developer:
+    Either make sure to include the plugins with the application (See feature
+    'disable_automatic_asset_installation'), or make them available in the
+    plugin directory.
+
+Possible fix for your problem:
+    Run `zellij setup --dump-plugins`, and optionally point it to your
+    'DATA DIR', visible in e.g. the output of `zellij setup --check`. Without
+    further arguments, it will use the default 'DATA DIR'.
+"
+    )]
+    BuiltinPluginMissing {
+        plugin_path: PathBuf,
+        plugin_dir: PathBuf,
+        #[source]
+        source: anyhow::Error,
+    },
+
+    #[error(
+        "It seems you tried to load the following builtin plugin:
+
+    Plugin name: '{plugin_path}'
+
+This is not a builtin plugin known to this version of zellij. If you were using
+a custom layout, please refer to the layout documentation at:
+
+    https://zellij.dev/documentation/creating-a-layout.html#plugin
+
+If you think this is a bug and the plugin is indeed an internal plugin, please
+open an issue on GitHub:
+
+    https://github.com/zellij-org/zellij/issues
+"
+    )]
+    BuiltinPluginNonexistent {
+        plugin_path: PathBuf,
+        #[source]
+        source: anyhow::Error,
+    },
 
     #[error("an error occured")]
     GenericError { source: anyhow::Error },