+++ zellij-utils/src/input/plugins.rs
@@ -9,6 +9,8 @@ use serde::{Deserialize, Serialize};
 use url::Url;
 
 use super::layout::{RunPlugin, RunPluginLocation};
+#[cfg(not(target_family = "wasm"))]
+use crate::consts::ASSET_MAP;
 pub use crate::data::PluginTag;
 use crate::errors::prelude::*;
 
@@ -74,7 +76,7 @@ impl Default for PluginsConfig {
 }
 
 /// Plugin metadata
-#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Deserialize, Serialize)]
 pub struct PluginConfig {
     /// Path of the plugin, see resolve_wasm_bytes for resolution semantics
     pub path: PathBuf,
@@ -87,16 +89,20 @@ pub struct PluginConfig {
 }
 
 impl PluginConfig {
-    /// Resolve wasm plugin bytes for the plugin path and given plugin directory. Attempts to first
-    /// resolve the plugin path as an absolute path, then adds a ".wasm" extension to the path and
-    /// resolves that, finally we use the plugin directory joined with the path with an appended
-    /// ".wasm" extension. So if our path is "tab-bar" and the given plugin dir is
-    /// "/home/bob/.zellij/plugins" the lookup chain will be this:
+    /// Resolve wasm plugin bytes for the plugin path and given plugin directory.
+    ///
+    /// If zellij was built without the 'disable_automatic_asset_installation' feature, builtin
+    /// plugins (Starting with 'zellij:' in the layout file) are loaded directly from the
+    /// binary-internal asset map. Otherwise:
+    ///
+    /// Attempts to first resolve the plugin path as an absolute path, then adds a ".wasm"
+    /// extension to the path and resolves that, finally we use the plugin directory joined with
+    /// the path with an appended ".wasm" extension. So if our path is "tab-bar" and the given
+    /// plugin dir is "/home/bob/.zellij/plugins" the lookup chain will be this:
     ///
     /// ```bash
     ///   /tab-bar
     ///   /tab-bar.wasm
-    ///   /home/bob/.zellij/plugins/tab-bar.wasm
     /// ```
     ///
     pub fn resolve_wasm_bytes(&self, plugin_dir: &Path) -> Result<Vec<u8>> {
@@ -110,9 +116,8 @@ impl PluginConfig {
             &plugin_dir.join(&self.path).with_extension("wasm"),
         ];
         // Throw out dupes, because it's confusing to read that zellij checked the same plugin
-        // location multiple times
+        // location multiple times. Do NOT sort the vector here, because it will break the lookup!
         let mut paths = paths_arr.to_vec();
-        paths.sort_unstable();
         paths.dedup();
 
         // This looks weird and usually we would handle errors like this differently, but in this
@@ -122,8 +127,32 @@ impl PluginConfig {
         // spell it out right here.
         let mut last_err: Result<Vec<u8>> = Err(anyhow!("failed to load plugin from disk"));
         for path in paths {
+            // Check if the plugin path matches an entry in the asset map. If so, load it directly
+            // from memory, don't bother with the disk.
+            #[cfg(not(target_family = "wasm"))]
+            if !cfg!(feature = "disable_automatic_asset_installation") && self.is_builtin() {
+                let asset_path = PathBuf::from("plugins").join(path);
+                if let Some(bytes) = ASSET_MAP.get(&asset_path) {
+                    log::debug!("Loaded plugin '{}' from internal assets", path.display());
+
+                    if plugin_dir.join(path).with_extension("wasm").exists() {
+                        log::info!(
+                            "Plugin '{}' exists in the 'PLUGIN DIR' at '{}' but is being ignored",
+                            path.display(),
+                            plugin_dir.display()
+                        );
+                    }
+
+                    return Ok(bytes.to_vec());
+                }
+            }
+
+            // Try to read from disk
             match fs::read(&path) {
-                Ok(val) => return Ok(val),
+                Ok(val) => {
+                    log::debug!("Loaded plugin '{}' from disk", path.display());
+                    return Ok(val);
+                },
                 Err(err) => {
                     last_err = last_err.with_context(|| err_context(err, &path));
                 },
@@ -131,6 +160,29 @@ impl PluginConfig {
         }
 
         // Not reached if a plugin is found!
+        #[cfg(not(target_family = "wasm"))]
+        if self.is_builtin() {
+            // Layout requested a builtin plugin that wasn't found
+            let plugin_path = self.path.with_extension("wasm");
+
+            if cfg!(feature = "disable_automatic_asset_installation")
+                && ASSET_MAP.contains_key(&PathBuf::from("plugins").join(&plugin_path))
+            {
+                return Err(ZellijError::BuiltinPluginMissing {
+                    plugin_path,
+                    plugin_dir: plugin_dir.to_owned(),
+                    source: last_err.unwrap_err(),
+                })
+                .context("failed to load a plugin");
+            } else {
+                return Err(ZellijError::BuiltinPluginNonexistent {
+                    plugin_path,
+                    source: last_err.unwrap_err(),
+                })
+                .context("failed to load a plugin");
+            }
+        }
+
         return last_err;
     }
 
@@ -143,10 +195,14 @@ impl PluginConfig {
             PluginType::Headless => {},
         }
     }
+
+    pub fn is_builtin(&self) -> bool {
+        matches!(self.location, RunPluginLocation::Zellij(_))
+    }
 }
 
 /// Type of the plugin. Defaults to Pane.
-#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
+#[derive(Clone, Debug, PartialEq, Deserialize, Serialize, Hash, Eq)]
 #[serde(rename_all = "kebab-case")]
 pub enum PluginType {
     // TODO: A plugin with output that's cloned across every pane in a tab, or across the entire