+++ zellij-utils/src/shared.rs
@@ -24,6 +24,11 @@ mod unix_only {
     }
 }
 
+#[cfg(not(unix))]
+pub fn set_permissions(_path: &std::path::Path, _mode: u32) -> std::io::Result<()> {
+    Ok(())
+}
+
 pub fn ansi_len(s: &str) -> usize {
     from_utf8(&strip(s).unwrap()).unwrap().width()
 }